And `-w` option to make files/directories writable before purging
When the backup is running as a non-root user the purging script may
fail due to permission issues when directories do not have write
permissions.

This change adds the `-w` flag so that the purging script will chmod
the directory to remove just before trying to remove it.  Tests were
updated to simulate this situation with a non-root user.